# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Doctors Scheduler",
      "version": "1.0.0"
    },
    "schemes": [
      "http",
      "https"
    ],
    "basePath": "/api/v1",
    "paths": {
      "/doctor/all": {
        "get": {
          "summary": "Get all doctors",
          "description": "Retrieves the list of doctors",
          "responses": {
            "200": {
              "description": "The list of doctors",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Doctor"
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete all Doctors",
          "description": "Deletes all doctors from the database",
          "responses": {
            "200": {
              "description": "The list of doctors that were deleted",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Doctor"
                }
              }
            }
          }
        }
      },
      "/doctor": {
        "post": {
          "summary": "Create a doctor",
          "parameters": [
            {
              "name": "doctor",
              "in": "body",
              "schema": {
                "$ref": "#/definitions/NewDoctor"
              }
            }
          ],
          "description": "Creates a single doctor",
          "consumes": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "The newly created doctor",
              "schema": {
                "$ref": "#/definitions/Doctor"
              }
            }
          }
        }
      },
      "/doctor/{id}": {
        "get": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Get single doctor",
          "description": "Get doctor by id",
          "responses": {
            "200": {
              "description": "Single doctor",
              "schema": {
                "$ref": "#/definitions/Doctor"
              }
            }
          }
        },
        "delete": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Delete a single doctor",
          "description": "Delete doctor by id",
          "responses": {
            "200": {
              "description": "Remaining Doctors",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Doctor"
                }
              }
            }
          }
        }
      },
      "/patient/all": {
        "get": {
          "summary": "Get all patients",
          "description": "Retrieves the list of patients",
          "responses": {
            "200": {
              "description": "The list of patients",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Patient"
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete all Patients",
          "description": "Deletes all patients from the database",
          "responses": {
            "200": {
              "description": "The empty list of patients after deletion",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Patient"
                }
              }
            }
          }
        }
      },
      "/patient": {
        "post": {
          "summary": "Create a patient",
          "parameters": [
            {
              "name": "patient",
              "in": "body",
              "schema": {
                "$ref": "#/definitions/NewPatient"
              }
            }
          ],
          "description": "Creates a single patient",
          "consumes": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "The newly created patient",
              "schema": {
                "$ref": "#/definitions/Patient"
              }
            }
          }
        }
      },
      "/patient/{id}": {
        "get": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Get single patient",
          "description": "Get patient by id",
          "responses": {
            "200": {
              "description": "Single patient",
              "schema": {
                "$ref": "#/definitions/Patient"
              }
            }
          }
        },
        "delete": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Delete a single patient",
          "description": "Delete patient by id",
          "responses": {
            "200": {
              "description": "Remaining Patients",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Patient"
                }
              }
            }
          }
        }
      },
      "/visits": {
        "get": {
          "summary": "Get all visits",
          "description": "Retrieves list of visits",
          "responses": {
            "200": {
              "description": "The list of visits",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Visit"
                }
              }
            }
          }
        }
      },
      "/visits/patient/{id}": {
        "get": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Get patient visits",
          "description": "Retrieves list of visits by patient",
          "responses": {
            "200": {
              "description": "The list of visits by patient",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Visit"
                }
              }
            }
          }
        }
      },
      "/hospital/all": {
        "get": {
          "summary": "Get all hospitals",
          "description": "Retrieves the list of hospitals",
          "responses": {
            "200": {
              "description": "The list of hospitals",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Hospital"
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete all hospitals",
          "description": "Deletes all hospitals from the database",
          "responses": {
            "200": {
              "description": "The empty list of hospitals after deletion",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Hospital"
                }
              }
            }
          }
        }
      },
      "/hospital": {
        "post": {
          "summary": "Create a hospital",
          "parameters": [
            {
              "name": "hospital",
              "in": "body",
              "schema": {
                "$ref": "#/definitions/NewHospital"
              }
            }
          ],
          "description": "Creates a single hospital",
          "consumes": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "The newly created hospital",
              "schema": {
                "$ref": "#/definitions/Hospital"
              }
            }
          }
        }
      },
      "/hospital/{id}": {
        "get": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Get single hospital",
          "description": "Get hospital by id",
          "responses": {
            "200": {
              "description": "Single hospital",
              "schema": {
                "$ref": "#/definitions/Hospital"
              }
            }
          }
        },
        "delete": {
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "type": "integer",
              "required": true
            }
          ],
          "summary": "Delete a single hospital",
          "description": "Delete hospital by id",
          "responses": {
            "200": {
              "description": "Remaining hospital",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Hospital"
                }
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Doctor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "NewDoctor": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "NewPatient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Visit": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "integer"
          },
          "patientId": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          }
        }
      },
      "Hospital": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer"
          },
          "doctors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "patients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "NewHospital": {
        "type": "object",
        "properties": {
          "doctors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "patients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }